add AppendEntries function

+++ src/raft/log.go
@@ -1,6 +1,6 @@
 package raft
 
 type LogEntry struct {
-	Term  int64
-	Index int64
+	Term  int
+	Index int
 }
